dashboard/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
jinja2
aiosqlite
python-multipart
